Dropping support for passing a single string for fields.
Passing a single string for the sort field helped me shape the
implementation, but it isn't required.
If people are calling the function with a programmatic list then
they'll always call it with a list, even if only one field is in it.
If people are hard-coding a call with a single field then it seems a
small trade-off to have them put it into brackets. You gain a
consistent API.